requirements.txt
fastapi
uvicorn
beanie
python-jose[cryptography]
python-multipart
starlette
filetype
motor
passlib[bcrypt]
python-dotenv
PyJWT
email-validator
boto3
python-magic
Mangum
pydantic
pydantic-settings
typing_extensions